refactor: shift ui colors upwards
All the greyscale/ui colors have been shifted up 1 position, with an
added new mantle in between the previous mantle and crust.

This includes (sub)text colors as well. A ui element with
maximum contrast (crust + text) should never #000 on #fff. This was a
jarring amount of contrast, and not ideal for human consumption.